Move VideoPlayerComponent to SwingSample file
The VideoPlayerComponent class was migrated from AvPlayer.kt to SwingSample.kt to better organize and align components with their respective functionalities. This change enhances code modularity and structure, making it easier to maintain.
